+++ src/Janda.Runtime.Assembly/AssemblyLogger.cs
@@ -1,12 +1,13 @@
 ﻿using Microsoft.Extensions.Logging;
 using System;
 using System.Linq;
+using System.Reflection;
 
 namespace Janda.Runtime
 {
     public class AssemblyLogger : IAssemblyLogger
     {
-        public static LogLevel LogLevel { get; set; } = LogLevel.Information;
+        public static LogLevel LogLevel { get; set; } = LogLevel.Debug;
         public static string NamespaceStartsWith { get; set; } = typeof(AssemblyLogger).Namespace?.Split('.').FirstOrDefault();
 
         public AssemblyLogger(ILogger<AssemblyLogger> logger)
@@ -14,12 +15,32 @@ namespace Janda.Runtime
             AssemblyLoaded((assembly) => { logger.Log(LogLevel, "Loaded {@Assembly}", new { assembly }); }, NamespaceStartsWith);
         }
 
-        public static void AssemblyLoaded(Action<dynamic> callback, string namespaceStartsWith = null)
+        private static dynamic GetAssemblyInfo(Assembly assembly)
+        {
+            return new 
+            { 
+                Name = assembly.GetAssemblyName(), 
+                Version = assembly.GetInformationalVersion() 
+            };
+        }
+
+        public static void AssemblyLoaded(Action<dynamic> callback)
         {
             AppDomain.CurrentDomain.AssemblyLoad += (object sender, AssemblyLoadEventArgs args) =>
             {
-                if (namespaceStartsWith == null || args.LoadedAssembly.FullName.StartsWith(namespaceStartsWith))
-                    callback(new { Name = args.LoadedAssembly.GetAssemblyName(), Version = args.LoadedAssembly.GetInformationalVersion() });
+                callback(GetAssemblyInfo(args.LoadedAssembly));
+            };
+        }
+
+        public static void AssemblyLoaded(Action<dynamic> callback, string namespaceStartsWith)
+        {
+            if (namespaceStartsWith == null)
+                throw new ArgumentNullException(namespaceStartsWith);
+
+            AppDomain.CurrentDomain.AssemblyLoad += (object sender, AssemblyLoadEventArgs args) =>
+            {
+                if (args.LoadedAssembly.FullName.StartsWith(namespaceStartsWith))
+                    callback(GetAssemblyInfo(args.LoadedAssembly));
             };
         }
     }