+++ src/Janda.Runtime.Assembly/AssemblyLoggerExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging;
 
 namespace Janda.Runtime
 {
@@ -8,5 +9,26 @@ namespace Janda.Runtime
         {
             return services.AddTransient<IAssemblyLogger, AssemblyLogger>();
         }
+
+        public static IServiceCollection AddAssemblyLogger(this IServiceCollection services, LogLevel logLevel)
+        {
+            AssemblyLogger.LogLevel = logLevel;
+            return services.AddAssemblyLogger();
+        }
+
+        public static IServiceCollection AddAssemblyLogger(this IServiceCollection services, string namespaceStartsWith)
+        {
+            AssemblyLogger.NamespaceStartsWith = namespaceStartsWith;
+            return services.AddAssemblyLogger();
+        }
+
+
+        public static IServiceCollection AddAssemblyLogger(this IServiceCollection services, LogLevel logLevel, string namespaceStartsWith)
+        {
+            AssemblyLogger.NamespaceStartsWith = namespaceStartsWith;
+            AssemblyLogger.LogLevel = logLevel;
+            return services.AddAssemblyLogger();
+        }
+
     }
 }