+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## 1.1.0 (2020-06-14)
+
+
+### Features
+
+* Add assembly extensions ([8a76be1](http://nas/matt/Janda.Runtime.Assembly/commit/8a76be1f397afd73885167f0f2c6993367df4867))
+* Add assembly logger ([fbf3f1f](http://nas/matt/Janda.Runtime.Assembly/commit/fbf3f1f4cddcf02520c973f857047b2462288dd3))
+* Add parameters to extension methods ([e4ff169](http://nas/matt/Janda.Runtime.Assembly/commit/e4ff169b887c8ee75918084c8e414dd499597bb7))
+
 ## 1.0.0 (2020-06-14)